+++ src/test/java/ortelius/testsuite/LoginTestSuite.java
@@ -1,4 +1,4 @@
-package ortelius;
+package ortelius.testsuite;
 
 import io.cucumber.junit.CucumberOptions;
 import net.serenitybdd.cucumber.CucumberWithSerenity;
@@ -7,6 +7,7 @@ import org.junit.runner.RunWith;
 @RunWith(CucumberWithSerenity.class)
 @CucumberOptions(
         plugin = {"pretty"},
+        glue = {"ortelius/stepdefinitions"},
         features = "src/test/resources/features/Login/Login.feature"
 )
 public class LoginTestSuite {}
