+++ src/test/resources/features/applications/DetailsTab.feature
