+++ src/test/resources/features/endpoints/DetailsTab.feature
